requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alibabacloud-credentials==0.3.2
alibabacloud-ecs20140526==3.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.2
alibabacloud-tea==0.3.5
alibabacloud-tea-openapi==0.3.8
alibabacloud-tea-util==0.3.11
alibabacloud-tea-xml==0.0.2
alibabacloud-vpc20160428==2.0.20
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
concurrent-log-handler==0.9.19
cryptography==37.0.4
decorator==5.1.1
Django==4.2.9
django-baton==3.0.2
django-cors-headers==3.7.0
django-debug-toolbar==3.5.0
django-json-widget==1.1.1
django-tidb==4.2.3
djangorestframework==3.14.0
dogpile.cache==1.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
frozenlist==1.4.1
future==0.18.3
idna==3.6
importlib-metadata==4.12.0
inflection==0.5.1
iso8601==2.1.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keystoneauth1==5.5.0
multidict==6.0.5
munch==4.0.0
mysqlclient==2.1.0
netifaces==0.11.0
openpyxl==3.1.2
openstacksdk==0.100.0
os-service-types==1.7.0
packaging==23.2
pbr==6.0.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==21.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
pyvmomi==7.0
PyYAML==6.0.1
requests==2.31.0
requestsexceptions==1.4.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
uWSGI==2.0.20
yarl==1.9.4
zipp==3.17.0
responses==0.24.1